edit on scraper and network
what has changed_ 1> The scraper is now a bit smarter, 2>introduction of website indentifier so the get_webmap fetches the correct website. As one user may have multiple projects with multiple websites. 3> The chart itself is now boxes with "Title" of the webpage instead of a circle for better identification. 4> The stickiness is still there